img/sim/requirements.txt
numpy==1.17.0
Pillow==6.1.0
